requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
Jinja2==3.1.4
requests==2.32.3
Pillow==10.4.0
PyPDF2==3.0.1
python-docx==1.1.2
xmltodict==0.13.0
openai==1.95.1
azure-identity==1.17.1
azure-mgmt-resource==23.1.1
azure-mgmt-authorization==4.0.0
zipfile36==0.1.3
python-dotenv==1.0.1
gunicorn==23.0.0
